+++ crates/backtest/src/python/mod.rs
@@ -38,6 +38,7 @@ pub fn backtest(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<crate::node::BacktestNode>()?;
     m.add_class::<engine::PyBacktestEngine>()?;
     m.add_class::<crate::modules::fx_rollover::InterestRateRecord>()?;
+    m.add_class::<crate::modules::fx_rollover::RolloverInterestCalculator>()?;
     m.add_class::<crate::modules::fx_rollover::FXRolloverInterestModule>()?;
     Ok(())
 }